feat: add increase node operator limit to sandbox module on Holesky
