+++ modules/blockChain/contractAddresses.ts
@@ -123,3 +123,7 @@ export const SDVTRegistry: ChainAddressMap = {
   [CHAINS.Goerli]: '0x6370FA71b9Fd83aFC4196ee189a0d348C90E93b0',
   [CHAINS.Holesky]: '0x11a93807078f8BB880c1BD0ee4C387537de4b4b6',
 }
+
+export const SandboxNodeOperatorsRegistry: ChainAddressMap = {
+  [CHAINS.Holesky]: '0xD6C2ce3BB8bea2832496Ac8b5144819719f343AC',
+}
