+++ modules/motions/types.ts
@@ -31,6 +31,8 @@ export const MotionTypeForms = {
   SDVTNodeOperatorRewardAddressesSet: 'SDVTNodeOperatorRewardAddressesSet',
   SDVTNodeOperatorNamesSet: 'SDVTNodeOperatorNamesSet',
   SDVTNodeOperatorManagerChange: 'SDVTNodeOperatorManagerChange',
+
+  SandboxNodeOperatorIncreaseLimit: 'SandboxNodeOperatorIncreaseLimit',
 } as const
 // intentionally
 // eslint-disable-next-line @typescript-eslint/no-redeclare
