+++ modules/motions/ui/MotionDescription/MotionDescription.tsx
@@ -4,7 +4,6 @@ import { useMotionCreatedEvent } from 'modules/motions/hooks/useMotionCreatedEve
 import { useContractEvmScript } from 'modules/motions/hooks/useContractEvmScript'
 
 import { DescLEGOTopUp } from './DescLEGO'
-import { DescNodeOperatorIncreaseLimit } from './DescNodeOperator'
 import {
   DescReferralPartnerAdd,
   DescReferralPartnerTopUp,
@@ -27,6 +26,7 @@ import {
   TopUpWithLimitsAbi,
   RemoveAllowedRecipientAbi,
   AddAllowedRecipientAbi,
+  EvmIncreaseNodeOperatorStakingLimitAbi,
 } from 'generated'
 import { Motion, MotionType } from 'modules/motions/types'
 import { EvmUnrecognized } from 'modules/motions/evmAddresses'
@@ -41,6 +41,7 @@ import { DescSDVTNodeOperatorRewardAddressesSet } from './DescSDVTNodeOperatorRe
 import { DescSDVTNodeOperatorNamesSet } from './DescSDVTNodeOperatorNamesSet'
 import { DescSDVTNodeOperatorsAdd } from './DescSDVTNodeOperatorsAdd'
 import { DescSDVTNodeOperatorManagersChange } from './DescSDVTNodeOperatorManagersChange'
+import { DescNodeOperatorIncreaseLimit } from './DescNodeOperatorLimitIncrease'
 
 type DescWithLimitsProps = NestProps<
   TopUpWithLimitsAbi['decodeEVMScriptCallData']
@@ -52,8 +53,19 @@ type DescAllowedRecipientAddProps = NestProps<
   AddAllowedRecipientAbi['decodeEVMScriptCallData']
 >
 
+type DescNodeOperatorIncreaseLimitProps = NestProps<
+  EvmIncreaseNodeOperatorStakingLimitAbi['decodeEVMScriptCallData']
+>
+
 const MOTION_DESCRIPTIONS = {
-  [MotionType.NodeOperatorIncreaseLimit]: DescNodeOperatorIncreaseLimit,
+  [MotionType.NodeOperatorIncreaseLimit]: (
+    props: DescNodeOperatorIncreaseLimitProps,
+  ) => (
+    <DescNodeOperatorIncreaseLimit
+      {...props}
+      motionType={MotionType.NodeOperatorIncreaseLimit}
+    />
+  ),
   [MotionType.LEGOTopUp]: DescLEGOTopUp,
   [MotionType.RewardProgramAdd]: DescRewardProgramAdd,
   [MotionType.RewardProgramTopUp]: DescRewardProgramTopUp,
@@ -207,6 +219,14 @@ const MOTION_DESCRIPTIONS = {
     DescSDVTTargetValidatorLimitsUpdate,
   [MotionType.SDVTNodeOperatorManagerChange]:
     DescSDVTNodeOperatorManagersChange,
+  [MotionType.SandboxNodeOperatorIncreaseLimit]: (
+    props: DescNodeOperatorIncreaseLimitProps,
+  ) => (
+    <DescNodeOperatorIncreaseLimit
+      {...props}
+      motionType={MotionType.SandboxNodeOperatorIncreaseLimit}
+    />
+  ),
 } as const
 
 type Props = {
