+++ modules/motions/utils/getNodeOperatorRegistryType.ts
@@ -1,0 +1,10 @@
+import {
+  IncreaseLimitMotionType,
+  INCREASE_LIMIT_MOTION_MAP,
+} from '../constants'
+
+export function getNodeOperatorRegistryType(
+  motionType: IncreaseLimitMotionType,
+) {
+  return INCREASE_LIMIT_MOTION_MAP[motionType].registryType
+}
